Fix: mis-used createfunction
function the createfunction make new function which corresponed current
method. Because dex2ir processes on a per-method basis, it sets global
elements func, basicblock, and builder. some mistakes was occured here,
it is not valid to call the createfunction for make new function.
because it will change current handled global objects.

so, to fix it. now it created function the CreateFunctionDelcare.
it will create new function such like CreateFunction, but it is not
change global objects.